refactor: move `ShadowType` to reuse in SUN entity
