+++ src/parser/entities/consts.ts
@@ -1,0 +1,4 @@
+export enum ShadowType {
+  RayTrace = 0,
+  ShadowMap = 1,
+}
